Update README.md
Thrilled to share the completion of my latest project: a Stopwatch application!

I've just wrapped up this project, which allows users to easily start, stop, and reset a timer, providing precise time tracking in a clean and responsive interface. This simple yet effective tool is built using HTML, CSS, and JavaScript.

This project was an excellent opportunity to sharpen my skills in:

-Frontend Web Development

-Responsive Design

-JavaScript Programming

I'm incredibly thankful for the chance to learn and contribute to impactful projects at Prodigy InfoTech. Eagerly anticipating more challenges and growth ahead!